Allow index.js to read command files and execute
Replaced if/else statements with a command to fetch the command, first-
checking if it is the right command, and returning if not, in the
directory and executes the interaction in the command file.

Added default ping command for testing

Added method to deploy commands using the JSON data from files in the
commands directory.

-WC